Open scene portal item
