+++ src/Android/Xamarin.Android/Samples/Scene/OpenScenePortalItem/OpenScenePortalItem.jpg
