+++ src/Android/Xamarin.Android/Samples/Scene/OpenScenePortalItem/readme.metadata.json
@@ -1,10 +1,10 @@
 {
     "category": "Map",
     "description": "Open a web scene from a portal item.",
-    "formal_name": "OpenScene",
+    "formal_name": "OpenScenePortalItem",
     "ignore": false,
     "images": [
-        "OpenScene.jpg"
+        "OpenScenePortalItem.jpg"
     ],
     "keywords": [
         "portal",
@@ -16,7 +16,7 @@
     },
     "offline_data": [],
     "redirect_from": [
-        "/net/latest/uwp/sample-code/openscene.htm"
+        "/net/latest/android/sample-code/openscene.htm"
     ],
     "relevant_apis": [
         "PortalItem",
@@ -24,8 +24,7 @@
         "SceneView"
     ],
     "snippets": [
-        "OpenScene.xaml",
-        "OpenScene.xaml.cs"
+        "OpenScenePortalItem.cs"
     ],
     "title": "Open a scene (portal item)"
 }
