+++ src/Forms/Shared/Samples/Scene/OpenScenePortalItem/OpenScenePortalItem.xaml
@@ -2,7 +2,7 @@
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:esriUI="clr-namespace:Esri.ArcGISRuntime.Xamarin.Forms;assembly=Esri.ArcGISRuntime.Xamarin.Forms"
-             x:Class="ArcGISRuntime.Samples.OpenScene.OpenScene">
+             x:Class="ArcGISRuntime.Samples.OpenScenePortalItem.OpenScenePortalItem">
     <Grid>
         <esriUI:SceneView x:Name="MySceneView" />
     </Grid>
