+++ src/UWP/ArcGISRuntime.UWP.Viewer/Samples/Scene/OpenScenePortalItem/OpenScenePortalItem.jpg
