+++ src/UWP/ArcGISRuntime.UWP.Viewer/Samples/Scene/OpenScenePortalItem/OpenScenePortalItem.xaml
@@ -1,5 +1,5 @@
 ﻿<UserControl
-    x:Class="ArcGISRuntime.UWP.Samples.OpenScene.OpenScene"
+    x:Class="ArcGISRuntime.UWP.Samples.OpenScenePortalItem.OpenScenePortalItem"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:esriUI="using:Esri.ArcGISRuntime.UI.Controls">
