+++ src/UWP/ArcGISRuntime.UWP.Viewer/Samples/Scene/OpenScenePortalItem/OpenScenePortalItem.xaml.cs
@@ -7,41 +7,32 @@
 // "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific 
 // language governing permissions and limitations under the License.
 
-using Android.App;
-using Android.OS;
-using Android.Widget;
 using Esri.ArcGISRuntime.Mapping;
 using Esri.ArcGISRuntime.Portal;
-using Esri.ArcGISRuntime.UI.Controls;
 using System;
+using Windows.UI.Popups;
 
-namespace ArcGISRuntime.Samples.OpenScene
+namespace ArcGISRuntime.UWP.Samples.OpenScenePortalItem
 {
-    [Activity (ConfigurationChanges=Android.Content.PM.ConfigChanges.Orientation | Android.Content.PM.ConfigChanges.ScreenSize)]
     [ArcGISRuntime.Samples.Shared.Attributes.Sample(
         name: "Open a scene (portal item)",
         category: "Map",
         description: "Open a web scene from a portal item.",
         instructions: "When the sample opens, it will automatically display the scene from ArcGIS Online. Pan and zoom to explore the scene.",
         tags: new[] { "portal", "scene", "web scene" })]
-    public class OpenScene : Activity
+    public partial class OpenScenePortalItem
     {
         // Hold the ID of the portal item, which is a web scene.
         private const string ItemId = "c6f90b19164c4283884361005faea852";
 
-        // Hold a reference to the scene view.
-        private SceneView _mySceneView;
-
-        protected override void OnCreate(Bundle bundle)
+        public OpenScenePortalItem()
         {
-            base.OnCreate(bundle);
+            InitializeComponent();
 
-            Title = "Open scene (Portal item)";
-
-            CreateLayout();
+            // Setup the control references and execute initialization.
             Initialize();
         }
-
+        
         private async void Initialize()
         {
             try
@@ -53,25 +44,12 @@ namespace ArcGISRuntime.Samples.OpenScene
                 PortalItem websceneItem = await PortalItem.CreateAsync(portal, ItemId);
 
                 // Create and show the scene.
-                _mySceneView.Scene = new Scene(websceneItem);
+                MySceneView.Scene = new Scene(websceneItem);
             }
             catch (Exception e)
             {
-                new AlertDialog.Builder(this).SetMessage(e.ToString()).SetTitle("Error").Show();
+                await new MessageDialog(e.ToString(), "Error").ShowAsync();
             }
-        }
-
-        private void CreateLayout()
-        {
-            // Create a new vertical layout for the app
-            LinearLayout layout = new LinearLayout(this) {Orientation = Orientation.Vertical};
-
-            // Add the map view to the layout
-            _mySceneView = new SceneView(this);
-            layout.AddView(_mySceneView);
-
-            // Show the layout in the app
-            SetContentView(layout);
         }
     }
 }
