+++ src/UWP/ArcGISRuntime.UWP.Viewer/Samples/Scene/OpenScenePortalItem/readme.metadata.json
@@ -1,10 +1,10 @@
 {
     "category": "Map",
     "description": "Open a web scene from a portal item.",
-    "formal_name": "OpenScene",
+    "formal_name": "OpenScenePortalItem",
     "ignore": false,
     "images": [
-        "OpenScene.jpg"
+        "OpenScenePortalItem.jpg"
     ],
     "keywords": [
         "portal",
@@ -16,7 +16,7 @@
     },
     "offline_data": [],
     "redirect_from": [
-        "/net/latest/android/sample-code/openscene.htm"
+        "/net/latest/uwp/sample-code/openscene.htm"
     ],
     "relevant_apis": [
         "PortalItem",
@@ -24,7 +24,8 @@
         "SceneView"
     ],
     "snippets": [
-        "OpenScene.cs"
+        "OpenScenePortalItem.xaml",
+        "OpenScenePortalItem.xaml.cs"
     ],
     "title": "Open a scene (portal item)"
 }
