+++ src/WPF/ArcGISRuntime.WPF.Viewer/Samples/Scene/OpenScenePortalItem/OpenScenePortalItem.jpg
