+++ src/iOS/Xamarin.iOS/Samples/Scene/OpenScenePortalItem/OpenScenePortalItem.jpg
